[chore] Makefile: add gencodeowners (#40073)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
This PR adds a target to the Makefile to generate the CODEOWNERS file
locally without a GitHub PAT.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #40072

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Added myself to a random receiver as a code owner and ran `make
gencodeowners`. It generated the `CODEOWNERS` file correctly without
requiring me to have a GitHub PAT (I do not have one set up).

<!--Describe the documentation added.-->
#### Documentation
Added reference in `CONTRIBUTING.md` and in the `check-codeowners`
GitHub Action error message.

<!--Please delete paragraphs that you did not use before submitting.-->